reformat announcements card
